requirements.txt
troposphere==1.8.1
pyyaml
awscli
boto
awacs
